restore mount.rs
when adding this code: mount_change(dest, MountPropagationFlags::PRIVATE) .context("make dest private BEFORE mount")?
in an attempt to solve detections caused mounts to stop working, it probably can be solved with fsmount or something else but I think it is not compatible with older kernels. I will try to solve it with YPatch.